fix: CodeQLエラー・警告を修正
## 修正内容

### エラー修正
- ContainmentMatcher.cs: 整数オーバーフロー回避（double キャスト）

### 警告修正
- CloudAIPerformanceTests.cs: 未使用変数をdiscardに変更
- TokenConsumptionTrackerTests.cs: 不要なアップキャストを削除
- FallbackOrchestratorTests.cs: CancellationTokenSourceにusing追加
- GeneralSettingsViewModel.cs: 冗長な条件式を簡素化
- FallbackOrchestrator.cs: 未使用変数をdiscardに変更
- EngineAccessController.cs: 未使用変数をdiscardに変更
- AggregatedChunksReadyEventHandler.cs: null参照演算子を追加

+++ Baketa.Infrastructure/Validation/ContainmentMatcher.cs
@@ -401,7 +401,8 @@ public sealed class ContainmentMatcher : IContainmentMatcher
         else if (b.Bottom < a.Top)
             dy = a.Top - b.Bottom;
 
-        return (int)Math.Sqrt(dx * dx + dy * dy);
+        // double にキャストしてから乗算（整数オーバーフロー回避）
+        return (int)Math.Sqrt((double)dx * dx + (double)dy * dy);
     }
 
     /// <summary>
